add session user role
