+++ app/counselor/dashboard/page.tsx
@@ -4,15 +4,18 @@ import * as React from 'react';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
 import NavbarCounselor from "@/components/Navbar_counselor";
+import { useSession } from "next-auth/react";
 
 
 export default function Counselor() {
+    const { data: session } = useSession();
+    var counselorName;
+    if (session) {
+        counselorName = session?.user?.name || '';
+    }
     return (
         <main className={styles.main}>
-            <NavbarCounselor header="Counselor Dashboard"/>      
-            <Typography align="center" component="h1">
-                Counselor Dashboard Placeholder
-            </Typography>  
+            <NavbarCounselor header = {counselorName || ''} />        
             <Toolbar />
 
         </main>
