+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@apollo/server": "^4.11.0",
     "@as-integrations/next": "^3.1.0",
+    "@auth/prisma-adapter": "^2.4.2",
     "@emotion/react": "^11.13.0",
     "@emotion/styled": "^11.13.0",
     "@graphql-yoga/node": "^3.9.1",
@@ -18,6 +19,7 @@
     "@mui/material": "^5.16.6",
     "@mui/styled-engine-sc": "^6.0.0-alpha.18",
     "@mui/x-date-pickers": "^7.14.0",
+    "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^5.19.1",
     "axios": "^1.7.4",
     "bcrypt": "^5.1.1",
